feat: improve regex support for version files
This commit adds support for regex in version file upgrades where if the regex is provided, Comet will match all the files with it and replace the matched string with first captured group and new version. If no regex is provided, Comet will look for old version and replace it with the newer one.